+++ R/options.R
@@ -1,7 +1,5 @@
 .onLoad <- function(libname, pkgname) {
 
-
-
   invisible()
 
 }
@@ -9,8 +7,7 @@
 .onAttach <- function(libname, pkgname){
 
   attempt <- try(googleAuthR::gar_attach_auto_auth("https://www.googleapis.com/auth/devstorage.full_control",
-                                    environment_var = "GCS_AUTH_FILE",
-                                    travis_environment_var = "TRAVIS_GCS_AUTH_FILE"))
+                                    environment_var = "GCS_AUTH_FILE"))
 
   if(inherits(attempt, "try-error")){
     warning("Problem using auto-authentication when loading from GCS_AUTH_FILE: \n", attempt, "